aichallenge/workspace/src/aichallenge_submit/vlm_planner/experiment/03_heron_nvila/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "heron-nvila"
version = "0.1.0"
description = "VLM-based trajectory planner for autonomous driving"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "transformers[torch]==4.57.1",
    "pillow",
]
